requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
